Revert "CR-760 simplify, so mock cases only loaded once"
This reverts commit 7b5ade933ac67cea5bb8289c80e3a519dc01f9d7.
