+++ src/test/java/uk/gov/ons/ctp/integration/contcencucumber/cucSteps/address/TestAddressEndpoints.java
@@ -22,9 +22,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.client.HttpClientErrorException;
 import org.springframework.web.util.UriComponentsBuilder;
-import uk.gov.ons.ctp.integration.contcencucumber.main.SpringIntegrationTest;
+import uk.gov.ons.ctp.integration.contcencucumber.cucSteps.TestBase;
 
-public class TestAddressEndpoints extends SpringIntegrationTest {
+public class TestAddressEndpoints extends TestBase {
 
   private static final Logger log = LoggerFactory.getLogger(TestAddressEndpoints.class);
   private AddressQueryResponseDTO addressQueryResponseDTO;
